add ServerSideTransport. add recipe for SSR, fixes #53 add recipe for persisted data, fixes #64
