+++ README.md
@@ -127,6 +127,29 @@ const transport = new HTTPTransport('/myEndpoint', {headers: {Authorization}});
 If you'd like to replace the global `cashay.transport`, you can call just call `cashay.create({transport: newTransport})`.
 This is useful if you use custom `fetchOptions` that include an authorization token and you need to change it.
 
+### ServerSideTransport
+
+```js
+new ServerSideTransport(graphQLHandler, errorHandler)
+```
+
+- *`graphQLHandler({query, variables})`: a Promise that takes a `query` and `variables` prop, and returns the output from
+a `graphql` call.
+- `errorHandler`: A custom function that takes any GraphQL `errors` and converts them into a single `error` for your Redux state
+
+Example:
+```js
+import {graphql} from 'graphql';
+import Schema from './rootSchema';
+import {ServerSideTransport} from 'cashay';
+const graphQLHandler = function({query, variables}) => {
+  return graphql(Schema, query, null, null, variables);
+}
+const transport = new ServerSideTransport(graphQLHandler);
+```
+This is useful if you use your server for both graphql and server-side rendering
+and don't want cashay to make an HTTP roundtrip to itself.
+
 ### Queries
 
 ```js
