+++ src/index.js
@@ -1,4 +1,4 @@
 export cashayReducer from './normalize/duck';
 export cashay from './Cashay';
-export HTTPTransport from './HTTPTransport';
+export HTTPTransport from './transports/HTTPTransport';
 export transformSchema from './schema/transformSchema';
