+++ src/transports/HTTPTransport.js
@@ -1,4 +1,5 @@
 import fetch from 'isomorphic-fetch'
+import defaultHandleErrors from './defaultHandleErrors';
 
 export default class HTTPTransport {
   constructor(uri = '/graphql', init = {}, handleErrors = defaultHandleErrors) {
@@ -35,25 +36,3 @@ export default class HTTPTransport {
 
   }
 }
-
-const tryParse = str => {
-  let obj;
-  try {
-    obj = JSON.parse(str);
-  } catch (e) {
-    return false;
-  }
-  return obj;
-};
-
-const defaultHandleErrors = (request, errors, duckField = '_error') => {
-  if (!errors) return;
-  // expect a human to put an end-user-safe message in the message field
-  const firstErrorMessage = errors[0].message;
-  if (!firstErrorMessage) return {errors};
-  const parsedError = tryParse(firstErrorMessage);
-  if (parsedError && parsedError.hasOwnProperty(duckField)) {
-    return parsedError;
-  }
-  return {errors};
-};
